New input music file system
I was doing the thing, where i ask people to put their audio files to play in folder, then i tested this on android, and conclusion of this is it doesn't work. So now my way is doing input where you could put link of file in internet to play, and file is getting loaded. And you could play it.